Task6/requirements.txt
Django==5.2.9
djangorestframework
django-phonenumber-field
phonenumbers
Pillow
